chore: 더미 기도카드 반영
