+++ src/components/todayPray/TodayPrayCardList.tsx
@@ -9,6 +9,7 @@ import useBaseStore from "@/stores/baseStore";
 import { useEffect } from "react";
 import { getISOTodayDate } from "@/lib/utils";
 import TodayPrayInviteCompletedItem from "./TodayPrayInviteCompletedItem";
+import DummyPrayCardUI from "../prayCard/DummyPrayCardUI";
 
 const TodayPrayCardList = () => {
   const setPrayCardCarouselApi = useBaseStore(
@@ -55,16 +56,24 @@ const TodayPrayCardList = () => {
     <Carousel setApi={setPrayCardCarouselApi} opts={{ startIndex: 1 }}>
       <CarouselContent>
         <CarouselItem className="basis-5/6"></CarouselItem>
-        {filterdGroupPrayCardList?.map((prayCard, index) => (
+        {filterdGroupPrayCardList?.map((prayCard) => (
           <CarouselItem key={prayCard.id} className="basis-5/6">
             <PrayCardUI
               prayCard={prayCard}
-              listLength={filterdGroupPrayCardList.length}
-              listIndex={index}
               eventOption={{ where: "PrayCardList" }}
             />
           </CarouselItem>
         ))}
+        {memberList.length == 1 && (
+          <CarouselItem className="basis-5/6">
+            <DummyPrayCardUI
+              profileImage="/images/avatar/avatar_1.png"
+              name="기도친구 1"
+              content={`가족 모두가 건강하고 행복하게 지낼 수 있도록 기도해주세요.\n\n특히 부모님의 건강이 항상 좋기를 바랍니다.\n\n또한, 형제자매들이 서로 화목하게 지내며, 각자의 삶에서 성공과 기쁨을 누릴 수 있기를 기도합니다.`}
+              dayOffset={2}
+            />
+          </CarouselItem>
+        )}
         {isPrayToday && (
           <CarouselItem className="basis-5/6">
             {memberList.length == 1 ? (
